smb encrypt = desired (opportunistic encryption)
default: default
default: enable negotiation of encryption but will not turn on data encryption globally or per share

off: disable

desired: enable negotiation and will turn on data encryption on sessions and share connections for those clients that support it.

required: enable negotiation and turn on data encryption on sessions and share connections. Clients that do not support encryption will be denied access to the server.